feat: Add migration to add runner and evaluator columns to LevaExperiments table
This commit adds a migration file to add the `runner_class` and `evaluator_classes` columns to the `leva_experiments` table in the database. These columns will be used to store the runner class and evaluator classes for each experiment. This change allows for better organization and management of experiments in the Leva application.

Related to recent code changes that added the `runner_class` and `evaluator_classes` columns to the `LevaExperiment` model and updated the `experiments_controller` to permit these columns in the `experiment_params`. Also aligns with recent repository commits that updated the schema file and added the migration file for the `leva_experiments` table.