Refactor Leva's `RunnerResult` to add runner_class attribute
This commit adds the `runner_class` attribute to the `RunnerResult` model in the Leva module. The `runner_class` attribute is used to store the name of the runner class associated with the result. This change includes modifications to the model, migration, and tests.

Related to recent code changes that added the `runner_class` attribute to the `RunnerResult` model and migration files.